Fixed 'openmsx_info {machines,extensions} <name>'
This info topic retrieves the meta data from a machine/extension configuration
file. Though it was still (only) looking for the 'hardwareconfig.xml' file. So
this was broken for configs that were converted to '<name>.xml'.

I factored out all code that loads configuration files to a single routine. So
now _all_ code paths should look for '<name>.xml' and if that is not found for
'<name>/hardwwareconfig.xml'. (Though the only bug was this info topic).