Made host->msx joystick mapping configurable
Before this patch this was how host joystick input was mapped to the emulated
msx joystick:
 - host axis 0 is mapped to MSX left/right
 - host axis 1 is mapped to MSX up/down
 - all other axis are ignored
 - all even numbered host buttons are mapped to MSX button A
 - all odd  numbered host buttons are mapped to MSX button B

This patch makes this mapping configurable, though the default mapping is still
as above.

Per host joystick there's now a setting 'joystick<N>_config' (with N=1,2,...).
The value of this setting can be something like:
  A {button0 button2} B {button1 button3} LEFT -axis0 RIGHT +axis0 UP -axis1 DOWN +axis1
So this is formatted like a Tcl dict (you can use the Tcl 'dict' function to
manipulate the value of this setting). The keys of this dict are the 6 possible
MSX inputs. The value for each key is a list of host joystick inputs. These
inputs can be: button<N>, +axis<N> or -axis<N>.

This schema does allow to map one host input to multiple MSX inputs. E.g.
  A {button0 button2} B {button1 button2}
In this example host button0 presses MSX buttonA and host button1->B. But host
button2 will simultaneously press MSX A+B.

There is no distinction between MSX action buttons and directions. So host
buttons and axis can freely be mapped to any MSX input.