Change default value of 'inputdelay' setting to 0.0
This is a follow-up on RELEASE_0_9_1-375-g7f70e53, see that commit more lots
more details. This patch effectively disables the EventDelay mechanism. I
didn't remove the code completely because in rare cases it might still be
useful and we will need something like this in the future when 'netplay' gets
implemented.