Fixed more users of 'machine_info slot'
The segment column in the openmsx-debugger was broken (thanks to BiFi for
reporting it).

This was caused by the change in revision RELEASE_0_9_1-396-g771cbae. That
revision made an incompatible change in the 'machine_info slot' command.
Usually we try to keep stuff backwards compatible, though in this specific case
the old behavior was simply wrong in some (rare) cases. Anyway, I though I
fixed all users of this function but apparently I was mistaken.

The debugger itself also uses this function. That one will be fixed in a
separate commit because it's in a different repo.