Fixed new ResampleHQ for 32-bit x86
The intrinsic _mm_cvtsi128_si64() only exists for 64-bit. So the previous patch
broke the 32-bit build on x86 machines. Luckily the workaround is not too
difficult. Though apparently it is difficult to get this asm stuff right on the
first try ;-)

The old asm code always went via the (old) MMX registers to store the 64-bit
result to memory. Though because this does require adding an extra (slow)
'emms' instruction I explicitly wanted to avoid that in the new code. Anyway,
this is not on the hot-code-path, so the performace point is mood, but the
new code may actually be faster even on 32-bit CPUs.