Rewrite drawNoiseLine asm routine using compiler intrinsics
Old code contained MMX, MMX-ext and SSE2 asm versions of this routine (for vc++
only SSE2, but one version for x86 and another for x64). The new code only has
a SSE2 version, but the same routine for gcc and vc++.

I don't expect any performance difference (it might be a tiny bit faster). The
purpose of this patch is to simplify the code (removes about 250 lines).

This patch also changes unsigned->uint32_t in some locations.