Use uint6_t/uint32_t as type for a 16/32bpp pixel
Before we used 'word' for 16bpp and 'unsigned' for 32bpp. The type 'word' was
first introduced in the MSX emulation context: for a Z80 a word really means 16
bits. But in other contexts a word may have a different meaning. In a lot of
places we use 'unsigned' when we want a 32 bit type. But the C++ standard
doesn't make any guarantees about the number of bits in that type. (Though in
all our current ports this assumption is true).

Anyway, since we switched to c++11 there are now standard typedef for types
with a specific bit width. So there's no excuse anymore not to use them. It
also makes the code easier to read IMHO. Now the video rendering code use
uint16_t/uint32_t for pixel types (of a specific bit depth) and 'unsigned' is
used to indicate simple unsigned values (I mean where the exact width doesn't
really matter). The type 'word' is no longer used in the video rendering code
(the VDP emulation code itself still uses it, but that's fine).