Replace scale_1on2 asm routines with intrinsics
Before, for gcc, there was a MMX routine for 16bpp and one for 32bpp. For vc++
there was only a MMX 32bpp routine, but separate for x86 and x64. Now there is
a single routine for gcc/vc++,16bpp/32bpp.

The routine was rewritten from MMX to SSE, so it can handle double the amount
of pixels with the same amount of instructions. Though I couldn't measure any
speed difference (positive nor negative), probably because the routine was
already limited by the memory bandwidth. (When benchmarking the MMX vs SSE
routine in isolation, with all data cached, the SSE version _is_ about 2x
faster). But when comparing to the plain c++ version the new routine does
perform better (certainly with clang, but also with gcc's auto-vectorized
version).

This patch also replaces 'long' or 'unsigned long' with 'size_t' (and sometimes
'prtdiff_t'). The idea was to have a type that's equally wide as a pointer,
both for 32 and 64 bit compilers. On gcc the type 'long' does fit this
description, but not on vc++. Before this patch this wasn't an issue because
(where it really mattered) 'long' was only used inside gcc-specific inline
assembly blocks.