Removed $(TCL_DEFS) from compiler command line
This fixes the problem I talked about in the commit message of rev 7300aef3.

The documentation on TCL_DEFS is very vague, but the following two posts seem
to indicate that using TCL_DEFS is not needed on the compiler command line. And
my tests show that it indeed works without (on my setup).

   https://groups.google.com/forum/?fromgroups#!topic/comp.lang.tcl/ykRtCqbKHw4
   http://wiki.tcl.tk/11822