Refactor buffer management in video scaling code
First some background information:

The job of the scaler (both software scaler and openGL hardware scalers) is to
transform the MSX graphics represented by a 'FrameSource' object to a (scaled)
viewable image. 'FrameSource' is an abstract class, concrete instances of this
class are:
- RawFrame: this more or less directly corresponds with the rendered output
  pixels of the VDP/V9990.
- DoubledFrame and DeinterlacedFrame: these combine one or two RawFrames to get
  a single interlaced/deinterlaced image.
- SuperImposedFrame and SuperImposedVideoFrame: these combine the output of two
  video sources (VDP+laserdisc or VDP+V9990) into a single image.

A scalers applies an algorithm to the input pixels to calculate the output.
Typically such an algorithm (e.g. HQ2x) require some 'context' to decide how to
transform the pixels. So instead of only looking at a single pixel at a time,
also the surrounding pixels are taken into account (both in the horizontal and
vertical direction). The image in a 'FrameSource' is stored line-based, so this
means a typical scaler algorithm will simultaneously look at several lines.

The V99x8 (and even more the V9990) can render images at different resolutions
(e.g. 256 or 512 pixels wide + borders). It is possible to have a screen-split
where multiple (vertical) regions of the screen are rendered using a different
(horizontal) resolution. Note that the vertical borders are internally modeled
as lines with a horizontal resolution of 1 pixel. So in a typical MSX screen
there will already be two such 'splits' (top and bottom vertical border).

Instead of teaching each scaler algorithm how to treat input lines of mixed
resolutions, we convert lines near the edges of such splits to a uniform
resolution. Such converted lines have to be (temporarily) stored in memory
'somewhere'.

Super imposing has a similar problem: instead of teaching all scaler algorithms
how to combine two input images into a single output, we first combine the two
images and then feed that combined image to the scaler. (The lines of) that
combined image also (temporarily) need to be stored in memory. (To make things
even more complex: the lines of the two input images can have a different
horizontal resolution).

Anyway, all the above tried to explain why, during scaling, we need extra
buffers. The next part explains how those buffers are managed.

--

Before this patch the interaction between the scaler (called 'sc') and the
FrameSource (called 'fs') more or less went like this:

 sc: Hey fs, give me line 10 but make sure it has a width of 640 pixels (first
     scale it if you have to).
 fs: Ok here it is. [I had to internally allocate buffer space to scale it.]
 sc: Thanks, now give me line 11, also width=640.
 fs: Here it is. [This was easy, I could directly pass the data from RawFrame.]
 sc: Thanks, now give me line 12, width=640.
 fs: Here it is. [I again need buffer space. I'm not sure sc is already done
     using the buffer I gave him at line 10, so to be safe I allocated
     additional buffer space.]
 ..
 sc: BTW fs, I'm done scaling. Feel free to reuse the internal buffers you had
     to allocate, if any.

Now the above scenario goes like this:

 sc: Hey fs, give me line 10, width=640. In case you need it here's a buffer
     you can use.
 fs: Ok, here's your line. [I had to use the provided buffer.]
 sc: Thanks, now give me line 11. In case you need it here's another buffer.
     [I still need the content of line 10, so I'm passing a 2nd buffer.]
 fs: Ok, here's your line. [I didn't need to use the buffer that was passed to
     me.]
 sc: Thanks, now give me line 12 and in case you need it here's a buffer.
     [I don't need the content of line 10 anymore, so I've reused that first
     buffer.]
 fs: Here it is. [I had to use the provided buffer.]
 ..
 sc: [Ok done. All the buffers were mine, so I don't need to tell fs I'm no
     longer using them.]

Or in other words, the responsibility to allocate/free buffer space is moved
from the FrameSource class to the user of that class. The user (=scaler) has
more detailed knowledge about the lifetime of the buffers, so it can more
easily (=sooner) reuse them. Typically the caller can allocate the buffers on
the stack, while before FrameSource was forced to allocate them on the heap.
This should all help to keep the buffers warm(er) in the CPU cache.

While scaling typical MSX software (depending on the scaler algorithm and
possibly screensplits) only a couple lines per frame benefit from this better
buffering. Though when using superimpose (laserdisc or video9000) all the lines
in the frame benefit.