Various cleanups (split from later patch)
(One of) the following patch(es) will change the buffer management of the video
rendering. Because it's a rather large patch I've split off all cleanups into
this patch. There should be almost no (or very small) functional changes in
this part.

When looked at in isolation, some of the changes in this patch seem useless
(like renaming a variable from 'buf0' to 'bufA'). But I wanted to get the
'core patch' as small as possible.