Replaced FrameSource::getLine() with getLineColor()
(Note: this talks about the getLine() overload that only takes a single
parameter, this is the function that returns an unscaled line.) All but one
users of this function only looked at the first pixel of the returned line.
This is exactly what the new getLineColor() functions does. The only other
user (in SuperImposedVideoFrame) could be implemented (slightly) more
efficient by using getLineInfo() instead of getLine().

In isolation this change has very little benefit, but it does make the next
patch a whole lot smaller.