Avoid buffer copies in Scaler1 and Scaler2
When no scaler effect is applied (scale_factor=1 or factor=2, blur=0,
scanline=0) and using the SDL renderer, we could in some cases avoid a copy of
the pixel line buffer. These scaling conditions are typical for our current
small-device ports (dingoo, Android).

The most extreme case is likely superimposing (laserdisc or video9000), in that
case (and the above conditions) we can avoid a copy on every line. I only did
one quick benchmark: factor=1 with laserdisc. This ran almost 10% faster (this
is for the total emulation time, the dominating cost is theora decoding, so the
scaling speedup itself is likely a lot bigger).

This speedup was made possible by the previous video scaling patch (the buffer
management refactoring).