Added 'interleave_black_frame'
See (updated) documentation for a description of this feature. Very briefly:
this will cause openMSX to render twice as many frames (PAL->100Hz,
NTSC->120Hz) and every other frame is respectively a normal or a black frame
(the MSX part anyway, the console and the other OSD elements remain unchanged).
This is (only) useful on a 100/120Hz LCD screen to reduce motion blur. Since I
don't have such a monitor, I couldn't really test this feature (except that it
looks ugly with a 60Hz host monitor refresh rate).

Thanks to Capitán Harlok for creating the initial implementation and doing all
the testing, I only tweaked/generalized the implementation a bit.