Initial version of the Super Swangi mapper.
Needed by the game with the same name (which is also known as Super
Altered Beast, as it's a port of Altered Beast). This is a rare Korean
game by Clover, 1990. Thanks to BiFi/hap for finding out how the mapper
works. The implementation is probably not complete, but it's better to
have a working mapper already and later (when more information becomes
available) fill in the details.