Align SectorBuffer / added AlignedBuffer
Because of the previous few patches, all code that manipulates disk sector
data, now use SectorBuffers. This patch makes SectorBuffers a properly aligned
structure. Many fields inside a SectorBuffer, e.g. the 32-bit filesize field of
a FAT directory entry, are now also naturually aligned. This allows to use the
aligned versions of the Endian::L16/32-like types (see DiskImageUtils). These
are faster on some systems. But also on systems that can do unaligned memory
accesses (like x86), having an aligned SectorBuffer has advantages because a
memset() or memcpy() on a known-to-be-aligned buffer can generate shorter and
faster code than the same operation on a raw byte array. And we fairly often
memcpy/memset SectorBuffers.

The IDE/SCSI code uses buffers that are bigger than one sector, the code also
frequently passes pointers to those buffers around. Instead of trying to
compose those buffers as multiple SectorBuffers, I created a AlignedBuffer
helper class. This behaves much like a byte-pointer but with the added
guarantee that the pointed to buffer is properly aligned (and this guarantee is
visible-to/checked-by the compiler).