Fix compilation error on Android
According to the linux man pages, both the st_mtime field in 'struct stat' and
the parameter of the localtime() function have the same type, namely 'time_t'.
However on Android st_mtime is 'unsigned long' and localtime takes a time_t
(which is 'long').

The cleanup patch committed earlier today changed the explictly named type
'type_t' of a temporary variable to 'auto'. Normally this is fine (and 'auto'
makes sure there are no accidental hidden type conversions). Though in case of
Android we must have a conversion, otherwise the code doesn't compile.