Fix typing '#' on spanish keyboard layout in the console
To type '#' on a spanish keyboard layout, you need to type <alt-gr>+<3>. On
some systems (windows) this generates a SDL event with modfifiers CTRL+ALT.
However the console key handling code ignores all key combination with CTRL
(except for some special ones like CTRL-A, CTRL-H, ...). So this means it
wasn't possible to type '#' in the console. It is possible in e.g MSX-BASIC.

This patch changes the code in CommandConsole:
- First all the 'special' keys combination are checked. For these we do look
  at the modifier field, e.g. SHIFT+PGUP, CTRL+A, ENTER (without modifier),
  ...
- Next we ignore the modifier field and rely on the unicode field. So even if
  typing some character implicitly adds the CTRL modifier, if it generates a
  valid unicode value, we process it like a normal key.