Big setting class refactoring
First an (over-)simplified historic overview of the openMSX settings:
- Initially we had simple setting classes that stored their specific value type
  (so int, bool, ... in the subclass instead of string in the base class). We
  choose for a policy based design with static polymorphism instead of dynamic
  polymorphism with virtual functions.
- Later we added Tcl scripting to openMSX, and the settings had to be kept
  in sync with Tcl variables. This requires to/from string conversions.
- Next we added per-machine and global settings. For convenience and backwards
  compatibility, the per-machine settings still have a representative in the
  global namespace. This is implemented via the concept of a proxy-setting.
- Over time we also introduced special cases like read-only settings, settings
  with a dynamically changing list of allowed values, settings with special
  default values, ...
Anyway, just to say that the setting classes have changed a lot over time and
that the original design, which is currently still present in some way, may not
be the best for the current needs.

This patch simplifies the internal implementation of the core setting
classes. The integration of the settings with Tcl and the rest of openMSX is
mostly left unchanged.

New design: instead of storing the setting value in the specific setting
classes (IntegerSetting stores an 'int', etc) the value is stored in the base
'Setting' class. The value is stored as a TclObject, so this still allows
efficient querying of the native type and avoids needless from/to string
conversions.

The 'check for allowed values' stuff has changed: before we did the basic
checks in the policy-classes (e.g. check that an IntegerSetting only receives
strings that can be converted to int). Extra checks were done via an optional
SettingChecker object (e.g. check that the string passed to the joystick config
setting has the required format). Now both checks are unified in a single
std::function callback (new in c++11). So no more inheritance to specialize the
behavior.

There's a new abstract base class called 'BaseSetting'. Subclasses are
'Setting' and 'ProxySetting'. This is much cleaner compared to the old
situation for the proxy settings: the 'BaseSetting' class has far fewer state,
so there is no duplicated information anymore.

The 'Setting' class is still abstract. Concrete subclasses are 'IntegerSetting',
'BooleanSetting', ... As already mentioned the value of the setting is now
stored as a TclObject in the 'Setting' base class. This means that many
operations can already be implemented at that level instead of delegating to
the concrete subclasses (either via virtual functions or via policy classes).

This is a relatively large change, so there might be some fallout.