Workaround pre-gcc-4.7.2 compiler bug
On my laptop, but not on my desktop, openMSX triggered an assert on exit with
the console open. I tracked it down to a bug in the compiler (actually a bug in
the standard library that ships with the compiler). On my latop I'm running
Kubuntu 12.04 which still has gcc-4.6.3 as the default compiler, on my desktop
I have gcc-4.7.3. Here's the gcc bug report:
  http://gcc.gnu.org/bugzilla/show_bug.cgi?id=54351

Unfortunately gcc-4.6 is still widely used, so it's worth adding a workaround
for this bug in the code. We actually already had a workaround for this same
bug in MSXMotherBoard, this patch adds a similar workaround in Reactor.