Don't use get() to test whether a unique_ptr is nullptr
In other words simplify:
    if (myUniquePtr.get()) myUniquePtr->someFunction();
to
    if (myUniquePtr) myUniquePtr->someFunction();

And similarly simplify:
    myBoolean = myUniquePtr.get() == nullptr;
to
    myBoolean = myUniquePtr == nullptr;

The same applies to shared_ptr and SDLSurfacePtr. These simplification brings
the syntax for using raw- and smart-pointers close together. The previous patch
already used this in a small way ('Reactor::display' is (depending on compiler
version) either a raw- or unique-ptri, but we use it with the same syntax).

Before we used c++11, all our unique_ptr's were auto_ptr's. And we had our own
implementation of shared_ptr. An auto_ptr does not allow the above syntax.
That's why our code still used '.get()' all over the place. C++11 added
explicit conversion operators. This feature allowed to implement the new
convert-to-bool behavior of unique_ptr. (It was already possible in c++98, but
not in an easy/safe way).

I extended SDLSurfacePtr with the same 'explicit conversion to bool operator'.
I'm not sure visual c++ already supports this. (It's supposed to be supported
in version "11.0 nov'12". Is this a beta version? Do we use it?). Anyway, if it
breaks it shouldn't be too hard to add a workaround.