Fix 'roll' behavior of layer B in P1 mode
In register R#18 the V9990 has 2 bits to influence how the vertical display
line counter 'rolls' when crossing a 256-line boundary. But apparently this
only applies to layer A, not to layer B in the P1 display mode (all other modes
only have one layer).

Thanks to Antonio Javier Estrada Villegas for reporting this bug and providing
an easy to use test case. This made fixing the problem quite easy.