Remove support for pre-2.0 openGL versions
IIRC early openMSX versions only required openGL-1.4. Later we optionally used
1.5 features. Then we added the SDLGL-PP renderer which requires openGL-2.0,
but at that point older openGL versions were still supported via the SDLGL
renderer. Later we dropped the SDLGL renderer, so from that point on openMSX
effectively only ran with openGL-2.0.

There still were a lot of openGL version checks in the code. But if we require
openGL-2.0, both at compile- and at run-time, then this can be simplified. So
that's what this patch does.

In the future we'd like to support openGL-ES-2.0 (for Android). This patch is a
very small step in that direction (first simplify the existing code).