Normalize value of BooleanSetting to "true" or "false"
In addition to "true" and "false", Tcl also accepts "on", "off", "yes", "no"
(capitalized or not) and any integer value (zero or not) as boolean values. Our
BooleanSetting class (before and after the setting class refactoring (rev
dffe5f98a)) accepts the same values. But one difference is that on querying the
current value, the old BooleanSetting would always return "true" or "false",
while the new version returns the same format that was set (as do normal Tcl
settings).

So although the new way is more Tcl-like, it broke for example our "toggle"
script. For example:
  set throttle on
  toggle throttle
  --> throttle has now the value "true", it should be "false"

Fixed by restoring the old behavior: accept any boolean input, but internally
normalize to "true" or "false".