[4/15] Keep content of softwaredb.xml in memory
This is the original idea that started this patch series. This patch alone
is responsible for the majority of the startup-speed-gain (see first commit
message of this series).

At startup openMSX parses the 'softwaredb.xml' file and keeps the result in
memory (in RomDatabase) as many RomInfo objects. These RomInfo objects contain
several string fields. So this whole process creates _a_lot_ of string's. Each
string requires a (small) memory allocation. The idea is now to keep the whole
xml file in memory (for the whole duration of openMSX) and instead of creating
many small string objects, use string_ref objects that only refer to the
original strings in the XML (our rapidxml parser changes xml entities like &lt;
in place, so we really can refer to the xml data block).

At first sight this may seem to require a whole lot more memory, because the
original xml file has to be kept constantly in memory. Surprisingly the extra
memory requirement is only about 200kB (the total size of the softwaredb.xml
file is about 650kB). I guess this is because each string allocation has some
memory overhead (memory allocations are typically rounded up to a multiple of
16 bytes), and those are all avoided now. But the most important thing is that
it's a lot faster.

After the remaining patches in this series the total openMSX memory usage will
drop again to about the same level as before this patch (+/- a few 10's of kB).

--

Side-note: softwaredb.xml has been slowly growing over time. So it is important
to parse it as fast as possible. If the database becomes really huge then it may
become beneficial to move to a 'real' database format (e.g. use sqlite). Though
ATM I think the current approach is still OK (and simpler).