[11/15] map<map<..>> -> vector<vector<..>> in EventDistributor
Similar to the previous patches, transform a map to a vector. Though it's a bit
different in this patch:

EventDistributor used a nested map:
  map<EventType, multimap<Priority, EventListener*>>

The first level is the type of the event, this is an enum value. We know that
all these (numeric) values are consecutive and we know that (almost) all will
actually be in use. So instead of a map or a sorted vector, we can as well use
a fixed-size vector directly indexed with the EventType enum.

The second level is sorted on priority. There are also only a very limited
number of priorities. But because this is the last level and it is a multimap,
we can't use the same trick to use Priority as an index in a vector. One
possibility would be to introduce a 3rd level that handles multiple listeners
per priority. But because only very few listeners actually use these priorities
and because simply iterating over all the listeners (of all priorities) is more
important (for performance), I choose to replace this second level with a
sorted vector of pair<Priority, EventListener*>.

In the previous patches there was a very clear boundary between the data
structure construction- and query-phase. That's not the case here:
EventListeners can be dynamically (un)registered. So instead of building a
vector and sorting it once, we now have to insert new EventListeners in the
correct position to keep it sorted. This is a less efficient operation on a
vector. Though usually the vector is small and EventListeners (de)registration
is a much less frequent operation than actual event delivery.

This patch also slightly changes the priority-blocking mechanism. An
EventListeners can block an event from listeners with a lower priority than
itself (e.g. the console can block a keyboard-event for the msx keyboard).
Before there was a bitmask for which priorities to block. Now it's only
possible to block listeners from a specific priority onward. But that's good
enough.