[14/15] map->vector in FilePool
FilePool keeps a map from sha1-values to filenames. It also needs the
reverse-map. At startup the content of the .filecache file is parsed into these
two maps. Later new entries are added (and occasionally removed) to/from these
maps.

So once constructed the maps are not completely read-only. Though they are only
modified occasionally and always when performing heavy file-IO (e.g. scanning
for a file with matching sha1). So the performance of the initial parsing of
.filecache is likely dominant. So I converted both maps to sorted vectors (and
thus take care to keep the vectors sorted when inserting new entries).

There's one more complication. These were the original maps:
  typedef multimap<Sha1Sum, pair<time_t, string>> Pool;
  typedef map<string, Pool::iterator> ReversePool;
So 'Pool' maps from sha1 to filename and 'ReversePool' maps from filename back
to a specific entry in 'Pool'. Note that 'Pool' is a multimap, so there can
be multiple files with the same sha1 (so a reverse-map from filename to sha1
(instead of iterator) does not uniquely identify the original entry).

One property of std::map iterators is that they remain valid for as long as the
element they refer to remains in the map. On the other hand std::vector
iterators become invalid on each insert/erase operation on that vector (because
the underlying data block may be reallocated or just because all elements with
higher index shift position). So if we replace map->vector we cannot continue
using iterators in 'ReversePool'. (Using indices instead of iterators solves
the reallocation, but not the shift problem).

This is the new mapping I choose:
  typedef vector<tuple<Sha1Sum, time_t, string>> Pool;
  typedef vector<pair<string_ref, Sha1Sum>> ReversePool;
Note that the reverse mapping uses a string_ref instead of a string. This
string_ref will refer to the string in Pool (so the full filename string
doesn't need to be in memory twice). Also note that the reverse mapping points
to a sha1-value (not an iterator) and a few paragraphs above we mentioned that
this is not a unique mapping. To work around this we now perform a reverse
lookup in two steps: First lookup the range of files with matching sha1. Next,
withing this range, search for the entry with matching filename. Usually there
will only be very few files with the same sha1. So reverse lookup is still a
fast operation.