[1/15] Minor cleanups + new utilities in stl.hh
This patch series started as a (single) patch to load softwaredb.xml faster.
Some of the ideas in that patch could be applied in other locations as well.
In the end there were changes all over the place. So I tried to split my
patch again in logical units.

The recurring idea in this patch series is to replace a std::map or std::set
with a sorted std::vector. This has the following properties:
- Because the vector is sorted, lookup can use binary search and is just like
  for map and set O(log(N)).
- Vector is a more compact structure in memory than map or set. It's also
  continuous in memory which is good for cache performance.
- Though vector usually has some excess capacity, so the actually allocated
  memory is not all really in use. Map and set OTOH only allocate nodes that
  are actually in use. In my measurements I found that the spare capacity in
  vector is, very roughly and for my specific tests, of the same magnitude as
  the extra memory overhead for map and set. So total memory usage is about the
  same, but the total number of allocated blocks is _a_lot_ less for vector.
- Vector doesn't perform well on insert/remove in the middle. For map/set
  this doesn't matter.

As we'll see later in this patch series, often the code follows the following
pattern:
- In a first phase, a data structure (before: a map or set, now: a vector) is
  filled. There are no queries on this structure yet.
- In a later phase there are queries (lookups) on this structure, without
  adding new or deleting existing entries from this structure. (Or only
  rarely).

This is an ideal pattern to use vector instead of map. First fill the vector,
next sort it and finally perform binary searches in it. I measured that filling
plus sorting a vector is faster than just filling the equivalent map/set. I
also measured that lookups (binary searches) on vector are faster than the
equivalent operation on map/set.

--

A few statistics (for the whole patch series, not just this patch).
Summary:
 - Startup performance seems to be improved by 6%-8%
 - Total memory usage remained about constant.

I measured starting openmsx with a Tcl script that (only) contains this line:
  after time 0.0001 exit

Average result of 30 runs of 'perf stat':
 - 8% less executed instructions (302M -> 278M)
 - 7%-8% less required CPU cycles (353M -> 325M)
  Note that #cycles > #instructions. This is not unusual for startup code:
  initialization code is only executed once with cold caches. When measuring
  inner loops in a benchmark this ratio is often the reverse.

Callgrind result:
 - about 6% faster (cycle estimation: 296M -> 279M)
 - less than 1/3 the number of calls to operator new (45208 -> 14806)

Massif result:
 - Heap memory usage, 'total' is the sum of 'useful' and 'wasted':
     total:  7633kB -> 7608kB
     useful: 7033kB -> 7323kB
     wasted:  600kB ->  285kB
 - So although there is logically more heap memory in use, there is a lot less
   memory wasted in overhead (heap administration, fragmentation, rounding up
   to multiple of 16 bytes, ...).
 - Likely the more logically-in-use heap memory is because of vectors that are
   not filled to full capacity (not a problem for (cache-)performance).
 - The callgrind results show that there are a lot less dynamic memory
   allocations. This explains the more efficient use of the heap (less waste).

--

This patch adds a new file 'src/utils/stl.hh' which contains a few utilities
that make working with sorted vectors easier. It also has some (very minor)
cleanups that didn't below to any specific patch later in this series. So this
patch itself doesn't make any functional changes yet.