[2/15] Tweaks in StringOp::split() and StringOp::join()
Both functions could split/join on a number of characters, but all callers of
these function only ever passed a single character. So adjusted these functions
to only accept one char (they are a tiny bit faster now). Other functions in
StringOp are overloaded to either take a single or a number of chars. We could
re-add the old overloads in the future (when needed).

The split() function returns sub-strings of the given string. Instead of
constructing a bunch of new (smaller) string objects we can just as well refer
to parts of the existing string. So return string_ref instead of string. This
saves on memory allocations. But it does limit the lifetime of the result to
the lifetime of the input string. For the existing callers this lifetime
limitation is not a problem.

And similarly the join() function now accepts a collection of string_ref's
instead of string's.