[5/15] Store romdb in a vector instead of a map
At startup openMSX parses softwaredb.xml and stores the result in memory. Later
entries can be looked up by sha1sum. This clearly matches the pattern described
in the first patch of this series: first create a data structure without
querying it, later query it without changing it.

So this patch changes the map into a vector. It's a bit more complicated
because there can be both a user and a system version of softwaredb that have
to be merged into a single structure. Duplicates between these two version have
to be filtered out (and duplicates within one db have to be reported). I think
I added enough comments in the code to document this process.

This patch results in a modest speedup but in a fairly significant reduction
in memory usage: a large part of the extra memory usage from the previous patch
has already gone away.