[6/15] Replace map->vector in RomInfo
RomInfo contained a map that associates mapper-names with internal mapper-ids.
There was also another map that associates the mapper-id with extra info about
the mapper (like description and block-size). Both maps again follow the
create-then-use pattern, so it's more efficient to use a (sorted) vector.

Parsing of 'softwaredb.xml' performs many mapper-name lookups. So this patch
results in a small measurable performance gain.