Avoid assert in HMMC/LMMC VDP commands
The LMMC, LMCM and HMMC VDP commands don't properly wait for VRAM access slots
(it's a TODO, but not a big problem because they are anyway limited by the CPU
speed). The LMMC and HMMC commands didn't update an internal 'time' variable
(LMCM did). Since access slots are not used, this variable is also not really
required. Except for some asserts .... so this patch does update 'time' just
enough so that the asserts do pass.

I triggered this assert in 'Magical Labyrinth Remix' in a HMMC command. There
are many other MSX programs that use HMMC that don't trigger this assert. I
didn't investigate in detail (because the problem is already fixed), but it
seemed that MLR started a HMMC command and then didn't send data for a very
long time (longer than the duration of one frame). And this is indeed a
required condition to trigger the bug.