[3/3] Lazy-Tcl: actually use this new feature
The previous patch implemented the on-demand loading feature, this patch will
actually make use of it.

As mentioned in the previous patch, scripts have to opt-in for on-demand
loading. To do this you have to do two things:
- Rename the script so that the filename has a leading '_' (this prevents the
  script from being executed immediately at startup).
- Register the name(s) of the proc(s) that the script will define. This is done
  via the new 'register_lazy' Tcl helper function. I've put all these calls in
  a new script 'lazy.tcl' (and because this name does not start with '_' this
  script is executed at startup).

So that's basically all what this patch does: it renames a whole lot of scripts
and for each such script it adds one command in 'lazy.tcl'. It's more
interesting to look at the list of scripts that are _not_ loaded on-demand:
   autoplug.tcl              uses 'after boot'
   create_user_dirs.tcl      needs to happen at startup
   keybindings.tcl           creates global keybindings
   lazy.tcl                  does registration
   load_icons.tcl            icons are loaded at startup
   mode.tcl                  possibly activates new mode at startup
   reverse.tcl               handles auto enable reverse
   screenshot.tcl            renames existing screenshot command
   session_management.tcl    possibly load session at startup
   tabbed_machine_view.tcl   triggers on each machine_switch
   umrcallback.tcl           set di_halt_callback