[1/3] Lazy-Tcl: move keybindings
Recently (and more in the past) we've spend some time on optimizing the openMSX
startup speed. On desktop machines openMSX starts fast enough, but on small
device like dingoo/android this is still important. Over time we've added more
and more Tcl scripts to the standard openMSX installation. (Combined with the
recent optimizations) this caused Tcl to take about 22% of the total startup
time. After this patch series that ratio drops to about 8%.

On startup we execute all .tcl files that are found in the openMSX (user or
system) 'share/scripts' subdirectory. Many of these scripts contain 'toys' or
'utilities' that are only occasionally useful and/or only for specific users.
This patch series will, as much as possible, load those scripts on-demand. For
example only execute the big 'osd_menu.tcl' script once the menu is actually
opened for the first time.

The mechanism for this on-demand-loading is discussed in the next patch. But I will
already give some performance statistics (for the full patch series):
 start openmsx with a Tcl script that does 'after time 0.0001 exit'
 perf stat cycles:       325M -> 289M  (-11%)
           instructions: 280M -> 243M  (-13%)
 valgrind Tcl relative: 22% -> 8%
              absolute: 62M -> 18M
 massif heap total:  7420kB -> 6547kB  (-12%)
             useful: 7136kB -> 6265kB  (-12%)
             wasted:  283kB ->  282kB  (- 0%)
So the time spend initializing Tcl gets roughly divided by a factor 3. This
results in about 12% faster startup with also 12% less heap memory usage.

--

This patch itself only moves some key bindings from the specific Tcl scripts
(e.g. osd_menu.tcl binds the MENU key) to one common 'keybindings.tcl' script.
The reason for this is that you want 'keybindings.tcl' to be executed always
while the specific scripts should (if possible) be loaded on-demand. Another
advantage is that this makes it clearer which global keybindings are actually
set by (all) the Tcl scripts.